refactor dropdown into separate component
